v0.0.6 (#5)
* refactor(instrumentation): Simplify setup with auto-instrumentations-web

Replaces the manual, conditional setup of individual OpenTelemetry instrumentations with the `getWebAutoInstrumentations` helper function.

This change simplifies the initialization logic by removing the need to manually import and instantiate each instrumentation. Configuration for enabling/disabling specific instrumentations is now passed directly as an options object.

Additionally, the `initBrowserInspector` function is now asynchronous, returning a Promise that resolves with the tracer provider upon successful initialization.

* chore(release): Bump version to 0.0.6

Increment the package version from 0.0.5 to 0.0.6.

This update prepares the package for a new release, incorporating the latest changes and fixes.